middleware: Move client instrumentation functions to grpc_instrumentation.go
